requirements.txt
Django==5.1.7
djangorestframework>=3.15,<4.0
django-filter>=24.2